Refactor CI configurations
* Add support of python 3.11
* Update linters
* Fix new linter errors
