+++ .pre-commit-config.yaml
@@ -3,7 +3,7 @@
 repos:
 -   repo: https://github.com/psf/black
     # If you update the version here, also update it in tox.ini (py*-pytestlatest-linters)
-    rev: 22.3.0
+    rev: 23.3.0
     hooks:
     - id: black
       args:
@@ -11,25 +11,30 @@ repos:
         - "--target-version=py38"
         - "--target-version=py39"
         - "--target-version=py310"
+        - "--target-version=py311"
 -   repo: https://github.com/pycqa/isort
-    rev: 5.10.1
+    rev: 5.12.0
     hooks:
       - id: isort
         name: isort (python)
 -   repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.1.0
+    rev: v4.4.0
     hooks:
     -   id: trailing-whitespace
     -   id: end-of-file-fixer
     -   id: check-yaml
     -   id: check-added-large-files
 -   repo: https://github.com/asottile/pyupgrade
-    rev: v2.31.1
+    rev: v3.4.0
     hooks:
     -   id: pyupgrade
         args: ["--py37-plus"]
 -   repo: https://github.com/pre-commit/mirrors-mypy
-    rev: v0.942
+    rev: v1.3.0
     hooks:
     -   id: mypy
-        additional_dependencies: [types-setuptools, types-certifi]
+        additional_dependencies: [pydantic, types-setuptools, types-certifi]
+-   repo: https://github.com/tox-dev/tox-ini-fmt
+    rev: "1.1.0"
+    hooks:
+      - id: tox-ini-fmt
