+++ pytest_bdd/compatibility/functools.py
@@ -1,0 +1,8 @@
+import sys
+
+if sys.version_info < (3, 8):
+    from singledispatchmethod import singledispatchmethod
+else:
+    from functools import singledispatchmethod
+
+__all__ = ["singledispatchmethod"]
