+++ setup.cfg
@@ -71,7 +71,7 @@ struct_bdd =
     hjson
     json5
     pyhocon
-    tomli
+    tomli;python_version<"3.11.0"
     PyYAML
     types-PyYAML
