+++ tox.ini
@@ -1,61 +1,67 @@
 [tox]
+envlist =
+    py311-pre-commit
+    py311-pytestlatest-linters
+    py311-pytest{625, 73, 72, 71, 70, latest}-mypy
+    py311-pytestlatets-allure-coverage
+    py311-pytest{625, 73, 72, 71, 70}-coverage
+    py311-pytestlatest-gherkin{24, latest}-xdist-coverage
+    py39-pytest{62, 61, 60, 54, 53, 52, 51, 50}-coverage
+    py{310, 39, 38, 37}-pytestlatest-mypy
+    py{39, 38, 37}-pytestlatest-coverage
 distshare = {homedir}/.tox/distshare
-envlist = py39-pre-commit
-          py310-pytestlatest-linters,
-          ; python 3.10 is only supported by pytest >= 6.2.5:
-          py310-pytestlatets-allure-coverage,
-          py310-pytest{625,70,71}-coverage,
-          ; the rest of pytest runs need to use an older python:
-          py39-pytest{50,51,52,53,54,60,61,62}-coverage,
-          py{37,38,39}-pytestlatest-coverage,
-          py310-pytestlatest-gherkin{24,latest}-xdist-coverage
 
 [testenv]
-setenv =
-    coverage: _PYTEST_CMD=coverage run --append -m pytest
-    xdist: _PYTEST_MORE_ARGS=-n3 -rfsxX
 deps =
-    .[test]
     .[struct_bdd]
-
+    .[test]
     allure: .[allure]
-    pytestlatest: pytest
-    pytest71: pytest~=7.1.0
-    pytest70: pytest~=7.0.0
-    pytest625: pytest>=6.2.5, <6.3
-    pytest62: pytest>=6.2.0, <6.2.5
-    pytest61: pytest~=6.1.0
-    pytest60: pytest~=6.0.0
-    pytest54: pytest~=5.4.0
-    pytest53: pytest~=5.3.0
-    pytest52: pytest~=5.2.0
-    pytest51: pytest~=5.1.0
-    pytest50: pytest~=5.0.0
+    coverage: coverage
     gherkin24: gherkin-official~=24.0.0
     gherkinlatest: gherkin-official
-
-    mypy:  {[testenv:mypy]deps}
-    coverage: coverage
+    mypy: {[testenv:mypy]deps}
+    pytest50: pytest~=5.0.0
+    pytest51: pytest~=5.1.0
+    pytest52: pytest~=5.2.0
+    pytest53: pytest~=5.3.0
+    pytest54: pytest~=5.4.0
+    pytest60: pytest~=6.0.0
+    pytest61: pytest~=6.1.0
+    pytest62: pytest>=6.2,<6.2.5
+    pytest625: pytest>=6.2.5,<6.3
+    pytest70: pytest~=7.0.0
+    pytest71: pytest~=7.1.0
+    pytest72: pytest~=7.2.0
+    pytest73: pytest~=7.3.0
+    pytestlatest: pytest
     xdist: pytest-xdist
-commands = {env:_PYTEST_CMD:pytest} {env:_PYTEST_MORE_ARGS:} {posargs:-vvl}
+setenv =
+    coverage: _PYTEST_CMD = coverage run --append -m pytest
+    xdist: _PYTEST_MORE_ARGS = -n3 -rfsxX
+commands =
+    {env:_PYTEST_CMD:pytest} {env:_PYTEST_MORE_ARGS:} {posargs:-vvl}
 
-[testenv:py39-pre-commit]
+[testenv:py311-pre-commit]
 skip_install = true
-deps = pre-commit
-commands = pre-commit run --all-files
+deps =
+    pre-commit
+commands =
+    pre-commit run --all-files
 
+[testenv:py311-pytestlatest-linters]
+deps =
+    black~=23.3.0
+commands =
+    black --check --verbose setup.py docs pytest_bdd tests
 
-[testenv:py310-pytestlatest-linters]
-deps = black~=22.3
-commands = black --target-version=py37 --target-version=py38 --target-version=py39 --target-version=py310 --check --verbose setup.py docs pytest_bdd tests
-
-[testenv:mypy]
+[testenv:py{37,38,39,310,311}-pytest{latest,625,70,71,72,73}-mypy]
 deps =
     mypy
+    pydantic
+    types-certifi
     types-setuptools
-
-[testenv:py37-pytestlatest-mypy]
-commands = mypy
+commands =
+    mypy
 
 [gh-actions]
 python =
@@ -63,3 +69,4 @@ python =
     3.8: py38
     3.9: py39
     3.10: py310
+    3.11: py311
